vrrp: extract socket lifecycle to instance_socket.go
Pure code-motion split of pkg/vrrp/instance.go (#5661). Move the
per-instance raw-socket lifecycle helpers into a cohesive sibling
file, package vrrp:

  openSocket, openSocketWith, closeSockets, closeSocketDescriptors,
  clearSocketRefs.

Declarations moved verbatim — no rename of exported symbols, no
logic/timing/const change, and no reorder of any goroutine, defer,
channel, or mutex operation (whole functions moved intact). Imports
pruned by goimports; gofmt clean. Builds green.
